Let user toggle between user page and default page upon login
